requirements.txt
python-dotenv==0.21.0
pyTelegramBotAPI==4.10.0
peewee==3.16.0